// playwright-e2e/models/ccd/ccd-events.ts
export type CCDEvent = {
  id: string;
  name: string;
  description?: string;
  order?: number;
};

type CCDEvents = {
  CREATE_CLAIM: CCDEvent;
  CREATE_CLAIM_AFTER_PAYMENT: CCDEvent;
  CREATE_CLAIM_SPEC: CCDEvent;
  CREATE_CLAIM_SPEC_AFTER_PAYMENT: CCDEvent;
  NOTIFY_DEFENDANT_OF_CLAIM: CCDEvent;
  NOTIFY_DEFENDANT_OF_CLAIM_DETAILS: CCDEvent;
  ACKNOWLEDGE_CLAIM: CCDEvent;
  ACKNOWLEDGEMENT_OF_SERVICE: CCDEvent;
  ADD_DEFENDANT_LITIGATION_FRIEND: CCDEvent;
  DEFENDANT_RESPONSE: CCDEvent;
  DEFENDANT_RESPONSE_SPEC: CCDEvent;
  CLAIMANT_RESPONSE: CCDEvent;
  WITHDRAW_CLAIM: CCDEvent;
  ADD_OR_AMEND_CLAIM_DOCUMENTS: CCDEvent;
  DISCONTINUE_CLAIM: CCDEvent;
  CASE_PROCEEDS_IN_CASEMAN: CCDEvent;
  RESUBMIT_CLAIM: CCDEvent;
  INFORM_AGREED_EXTENSION_DATE: CCDEvent;
  INFORM_AGREED_EXTENSION_DATE_SPEC: CCDEvent;
  AMEND_PARTY_DETAILS: CCDEvent;
  ADD_CASE_NOTE: CCDEvent;
  CHANGE_SOLICITOR_EMAIL: CCDEvent;
  ENTER_BREATHING_SPACE_SPEC: CCDEvent;
  LIFT_BREATHING_SPACE_SPEC: CCDEvent;
  DEFAULT_JUDGEMENT: CCDEvent;
  DEFAULT_JUDGEMENT_SPEC: CCDEvent;
  CREATE_SDO: CCDEvent;
  CLAIMANT_RESPONSE_SPEC: CCDEvent;
  REFER_TO_JUDGE: CCDEvent;
  INITIATE_GENERAL_APPLICATION: CCDEvent;
  REQUEST_JUDGEMENT_ADMISSION_SPEC: CCDEvent;
  MANAGE_CONTACT_INFORMATION: CCDEvent;
  MANAGE_DOCUMENTS: CCDEvent;
  RESET_PIN: CCDEvent;
  CREATE_CASE_FLAGS: CCDEvent;
  MANAGE_CASE_FLAGS: CCDEvent;
};

export default CCDEvents;